Linear feature effects on lambda

Linear feature restoration (i.e. forest regeneration) and linear feature deactivation (e.g. blocking entrances) effects on lambda.

wildlift_linear(tmax=20, pop.start=100,
    area=10000, lin=0, seism=0, young=0,
    cost=12, yr_deact=5, yr_restor=15)

Arguments

tmax
positive integer, number of years to forecast after initial year 0.

pop.start
positive integer, initial population size in year 0.

area
positive numeric, area of range in km^2.

lin
non-negative numeric, linear features (total length in km) including seismic lines (to be restored or deactivated) and permanent linear featured (roads etc.).

seism
non-negative numeric, seismic lines (total length in km) to be restored or deactivated.

young
numeric between 0 and 100, percent of young (<30 yrs) forest.

cost
non-negative numeric, cost of management action (in canadian dollars) per km linear feature restored/deactivated.

yr_deact
non-negative numeric, years it takes for deactivation to reach a level where it begins to affect lambda.

yr_restor
non-negative numeric, years it takes for restoration to reach a level where it begins to affect lambda.

Details

Lambda is calculated as 1.0184 - 0.0234 * (lin / area) - 0.0021 * young.

Value

A list with population table and costs.

See also

Examples

## nothing to restore/deactivate
wildlift_linear()

#> $costdeact
#> [1] 0
#> 
#> $costrestor
#> [1] 0
#> 
#> $pop
#>       year  N0  N1 Ndeact Nrestor lddeact ldrestor young   lam0   lam1 lamdeact
#>  [1,]    0 100 100    100     100       0        0     0 1.0184 1.0184   1.0184
#>  [2,]    2 101 101    101     101       0        0     0 1.0184 1.0184   1.0184
#>  [3,]    3 102 102    102     102       0        0     0 1.0184 1.0184   1.0184
#>  [4,]    4 103 103    103     103       0        0     0 1.0184 1.0184   1.0184
#>  [5,]    5 104 104    104     104       0        0     0 1.0184 1.0184   1.0184
#>  [6,]    6 105 105    105     105       0        0     0 1.0184 1.0184   1.0184
#>  [7,]    7 106 106    106     106       0        0     0 1.0184 1.0184   1.0184
#>  [8,]    8 107 107    107     107       0        0     0 1.0184 1.0184   1.0184
#>  [9,]    9 108 108    108     108       0        0     0 1.0184 1.0184   1.0184
#> [10,]   10 109 109    109     109       0        0     0 1.0184 1.0184   1.0184
#> [11,]   11 111 111    111     111       0        0     0 1.0184 1.0184   1.0184
#> [12,]   12 113 113    113     113       0        0     0 1.0184 1.0184   1.0184
#> [13,]   13 115 115    115     115       0        0     0 1.0184 1.0184   1.0184
#> [14,]   14 117 117    117     117       0        0     0 1.0184 1.0184   1.0184
#> [15,]   15 119 119    119     119       0        0     0 1.0184 1.0184   1.0184
#> [16,]   16 121 121    121     121       0        0     0 1.0184 1.0184   1.0184
#> [17,]   17 123 123    123     123       0        0     0 1.0184 1.0184   1.0184
#> [18,]   18 125 125    125     125       0        0     0 1.0184 1.0184   1.0184
#> [19,]   19 127 127    127     127       0        0     0 1.0184 1.0184   1.0184
#> [20,]   20 129 129    129     129       0        0     0 1.0184 1.0184   1.0184
#> [21,]   21 131 131    131     131       0        0     0 1.0184 1.0184   1.0184
#>       lamrestor
#>  [1,]    1.0184
#>  [2,]    1.0184
#>  [3,]    1.0184
#>  [4,]    1.0184
#>  [5,]    1.0184
#>  [6,]    1.0184
#>  [7,]    1.0184
#>  [8,]    1.0184
#>  [9,]    1.0184
#> [10,]    1.0184
#> [11,]    1.0184
#> [12,]    1.0184
#> [13,]    1.0184
#> [14,]    1.0184
#> [15,]    1.0184
#> [16,]    1.0184
#> [17,]    1.0184
#> [18,]    1.0184
#> [19,]    1.0184
#> [20,]    1.0184
#> [21,]    1.0184
#> 
#> $call
#> wildlift_linear()
#> 
#> attr(,"class")
#> [1] "wildlift_linear"

## all linear restored/deactivated
wildlift_linear(lin=1000, seism=1000, young=10)

#> $costdeact
#> [1] 12
#> 
#> $costrestor
#> [1] 12
#> 
#> $pop
#>       year  N0  N1 Ndeact Nrestor lddeact    ldrestor young   lam0    lam1
#>  [1,]    0 100 100    100     100    0.10 0.100000000    10 0.9974 0.99506
#>  [2,]    2  99  99     99      99    0.08 0.093333333    10 0.9974 0.99506
#>  [3,]    3  98  98     98      98    0.06 0.086666667    10 0.9974 0.99506
#>  [4,]    4  97  97     97      97    0.04 0.080000000    10 0.9974 0.99506
#>  [5,]    5  96  96     96      96    0.02 0.073333333    10 0.9974 0.99506
#>  [6,]    6  95  95     95      95    0.00 0.066666667    10 0.9974 0.99506
#>  [7,]    7  94  94     94      94    0.00 0.060000000    10 0.9974 0.99506
#>  [8,]    8  93  93     93      93    0.00 0.053333333    10 0.9974 0.99506
#>  [9,]    9  92  92     92      92    0.00 0.046666667    10 0.9974 0.99506
#> [10,]   10  91  91     91      91    0.00 0.040000000    10 0.9974 0.99506
#> [11,]   11  90  90     90      90    0.00 0.033333333    10 0.9974 0.99506
#> [12,]   12  89  89     89      89    0.00 0.026666667    10 0.9974 0.99506
#> [13,]   13  88  88     88      88    0.00 0.020000000    10 0.9974 0.99506
#> [14,]   14  87  87     87      87    0.00 0.013333333    10 0.9974 0.99506
#> [15,]   15  86  86     86      86    0.00 0.006666667    10 0.9974 0.99506
#> [16,]   16  85  85     85      85    0.00 0.000000000    10 0.9974 0.99506
#> [17,]   17  84  84     84      84    0.00 0.000000000    10 0.9974 0.99506
#> [18,]   18  83  83     83      83    0.00 0.000000000    10 0.9974 0.99506
#> [19,]   19  82  82     82      82    0.00 0.000000000    10 0.9974 0.99506
#> [20,]   20  81  81     81      81    0.00 0.000000000    10 0.9974 0.99506
#> [21,]   21  80  80     80      80    0.00 0.000000000    10 0.9974 0.99506
#>       lamdeact lamrestor
#>  [1,] 0.995060  0.995060
#>  [2,] 0.995528  0.995216
#>  [3,] 0.995996  0.995372
#>  [4,] 0.996464  0.995528
#>  [5,] 0.996932  0.995684
#>  [6,] 0.997400  0.995840
#>  [7,] 0.997400  0.995996
#>  [8,] 0.997400  0.996152
#>  [9,] 0.997400  0.996308
#> [10,] 0.997400  0.996464
#> [11,] 0.997400  0.996620
#> [12,] 0.997400  0.996776
#> [13,] 0.997400  0.996932
#> [14,] 0.997400  0.997088
#> [15,] 0.997400  0.997244
#> [16,] 0.997400  0.997400
#> [17,] 0.997400  0.997400
#> [18,] 0.997400  0.997400
#> [19,] 0.997400  0.997400
#> [20,] 0.997400  0.997400
#> [21,] 0.997400  0.997400
#> 
#> $call
#> wildlift_linear(lin = 1000, seism = 1000, young = 10)
#> 
#> attr(,"class")
#> [1] "wildlift_linear"

## half of linear is to be restored/deactivated
wildlift_linear(lin=1000, seism=500, young=10)

#> $costdeact
#> [1] 6
#> 
#> $costrestor
#> [1] 6
#> 
#> $pop
#>       year  N0  N1 Ndeact Nrestor lddeact   ldrestor young   lam0    lam1
#>  [1,]    0 100 100    100     100    0.10 0.10000000    10 0.9974 0.99506
#>  [2,]    2  99  99     99      99    0.08 0.09333333    10 0.9974 0.99506
#>  [3,]    3  98  98     98      98    0.06 0.08666667    10 0.9974 0.99506
#>  [4,]    4  97  97     97      97    0.05 0.08000000    10 0.9974 0.99506
#>  [5,]    5  96  96     96      96    0.05 0.07333333    10 0.9974 0.99506
#>  [6,]    6  95  95     95      95    0.05 0.06666667    10 0.9974 0.99506
#>  [7,]    7  94  94     94      94    0.05 0.06000000    10 0.9974 0.99506
#>  [8,]    8  93  93     93      93    0.05 0.05333333    10 0.9974 0.99506
#>  [9,]    9  92  92     92      92    0.05 0.05000000    10 0.9974 0.99506
#> [10,]   10  91  91     91      91    0.05 0.05000000    10 0.9974 0.99506
#> [11,]   11  90  90     90      90    0.05 0.05000000    10 0.9974 0.99506
#> [12,]   12  89  89     89      89    0.05 0.05000000    10 0.9974 0.99506
#> [13,]   13  88  88     88      88    0.05 0.05000000    10 0.9974 0.99506
#> [14,]   14  87  87     87      87    0.05 0.05000000    10 0.9974 0.99506
#> [15,]   15  86  86     86      86    0.05 0.05000000    10 0.9974 0.99506
#> [16,]   16  85  85     85      85    0.05 0.05000000    10 0.9974 0.99506
#> [17,]   17  84  84     84      84    0.05 0.05000000    10 0.9974 0.99506
#> [18,]   18  83  83     83      83    0.05 0.05000000    10 0.9974 0.99506
#> [19,]   19  82  82     82      82    0.05 0.05000000    10 0.9974 0.99506
#> [20,]   20  81  81     81      81    0.05 0.05000000    10 0.9974 0.99506
#> [21,]   21  80  80     80      80    0.05 0.05000000    10 0.9974 0.99506
#>       lamdeact lamrestor
#>  [1,] 0.995060  0.995060
#>  [2,] 0.995528  0.995216
#>  [3,] 0.995996  0.995372
#>  [4,] 0.996230  0.995528
#>  [5,] 0.996230  0.995684
#>  [6,] 0.996230  0.995840
#>  [7,] 0.996230  0.995996
#>  [8,] 0.996230  0.996152
#>  [9,] 0.996230  0.996230
#> [10,] 0.996230  0.996230
#> [11,] 0.996230  0.996230
#> [12,] 0.996230  0.996230
#> [13,] 0.996230  0.996230
#> [14,] 0.996230  0.996230
#> [15,] 0.996230  0.996230
#> [16,] 0.996230  0.996230
#> [17,] 0.996230  0.996230
#> [18,] 0.996230  0.996230
#> [19,] 0.996230  0.996230
#> [20,] 0.996230  0.996230
#> [21,] 0.996230  0.996230
#> 
#> $call
#> wildlift_linear(lin = 1000, seism = 500, young = 10)
#> 
#> attr(,"class")
#> [1] "wildlift_linear"